refactor proposal memo lineage router
